fix compareName - if min name is 2 then it should be fully equal, if 3 then only 1 diff is valid, change diff from uint to int

+++ state/state.go
@@ -439,7 +439,7 @@ func (root *Root) FindFamily(name string) *Family {
 	}
 
 	source := []rune(name)
-	min := uint(len(source))
+	min := len(source)
 
 	for key, f := range root.Families {
 		diff := compareNames(source, []rune(key))
